autonomous_brewing/central_server/api_backend/requirements.txt
django
djangorestframework
redis
pymongo
django-cors-headers
